Various project updates and cleanups

+++ src/main/kotlin/com/curiouscreature/kotlin/math/Matrix.kt
@@ -462,8 +462,8 @@ fun translation(m: Mat4) = translation(m.translation)
 fun rotation(m: Mat4) = Mat4(normalize(m.right), normalize(m.up), normalize(m.forward))
 fun rotation(d: Float3): Mat4 {
     val r = transform(d, ::radians)
-    val c = transform(r, { x -> cos(x.toDouble()).toFloat()})
-    val s = transform(r, { x -> sin(x.toDouble()).toFloat()})
+    val c = transform(r, { x -> cos(x) })
+    val s = transform(r, { x -> sin(x) })
 
     return Mat4.of(
              c.y * c.z, -c.x * s.z + s.x * s.y * c.z,  s.x * s.z + c.x * s.y * c.z, 0.0f,
